REFACTOR: (MOISAM-248) log 관련 클래스는 log 패키지에서 관리한다
