+++ src/main/java/com/meetup/server/admin/application/AdminService.java
@@ -9,10 +9,10 @@ import com.meetup.server.admin.implement.AdminValidator;
 import com.meetup.server.admin.implement.AdminWriter;
 import com.meetup.server.event.domain.Event;
 import com.meetup.server.event.implement.EventReader;
+import com.meetup.server.log.implement.LogUserLoginReader;
 import com.meetup.server.startpoint.implement.StartPointReader;
 import com.meetup.server.startpoint.infrastructure.querydsl.projection.ParticipantCount;
 import com.meetup.server.user.domain.User;
-import com.meetup.server.user.implement.LogUserLoginReader;
 import com.meetup.server.user.implement.UserReader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
