+++ src/main/java/com/meetup/server/auth/support/handler/OAuth2LoginSuccessHandler.java
@@ -3,8 +3,8 @@ package com.meetup.server.auth.support.handler;
 import com.meetup.server.auth.dto.CustomOAuth2User;
 import com.meetup.server.auth.support.CookieUtil;
 import com.meetup.server.global.support.jwt.JwtTokenProvider;
-import com.meetup.server.user.domain.type.LoginStatus;
-import com.meetup.server.user.implement.LogUserLoginWriter;
+import com.meetup.server.log.domain.type.LoginStatus;
+import com.meetup.server.log.implement.LogUserLoginWriter;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
